+++ medirect-currency-exchange.Domain/Models/Customer.cs
@@ -2,14 +2,18 @@
 {
 	public class Customer
 	{
-		public double Id { get; set; }
+		public long Id { get; set; }
+		
 		public string Name { get; set; }
+		
 		public string Surname { get; set; }
+		
 		public string IdCard { get; set; }
+		
 		public DateTime DOB { get; set; }
+		
 		public string Email { get; set; }
 
-		
 		public IEnumerable<CustomerWallet> Wallets { get; set; }
 	}
 }
