+++ medirect-currency-exchange.Domain/Models/CustomerWallet.cs
@@ -4,11 +4,11 @@ namespace medirect_currency_exchange.Domain.Models
 {
 	public class CustomerWallet
 	{
-		public double CustomerId { get; set; }
+		public long CustomerId { get; set; }
 
 		public string CurrencyCode { get; set; }
+		
 		[Column(TypeName = "decimal(18,5)")]
-
 		public decimal Amount { get; set; }
 
 		public DateTime LastModified { get; set; }
